src/services/image-recognition/requirements.txt
numpy
tensorflow
Pillow
Flask
flasgger